requirements.txt
Django~=5.1.7

pillow~=11.1.0
